+++ CHANGELOG.md
@@ -7,6 +7,10 @@ Changelog
 
 - Added SMASH server configuration options ([PR 2259](https://github.com/input-output-hk/daedalus/pull/2259))
 
+### Chores
+
+- Disabled application menu navigation before the "Terms of use" have been accepted ([PR 2304](https://github.com/input-output-hk/daedalus/pull/2304))
+
 ## 3.2.0-FC1
 
 ### Features